Broke our implant.go into two files.
Moving by functionality seems sane.
